backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hfc-scoring-engine"
version = "1.0.0"
description = "HFC Scoring Engine - Lead qualification and scoring API"
requires-python = ">=3.9"
dependencies = [
    "Django>=3.2,<4.0",
    "djangorestframework>=3.12,<4.0",
    "django-cors-headers>=3.10,<4.0",
    "gunicorn>=20.1,<21.0",
    "psycopg2-binary>=2.9,<3.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
